Fuzzer: improve performance
* Use persistent mode for fuzzing - gives a 2-3x speedup
* use nullwriter for AST dump and omit diagnostic rendering
* reorganize file locations and remove debug-helper - We can just use the compiler
	itself when trying to debug crashes
* manually remove fuzz input source so we don't need to dupe it
