+++ build.zig
@@ -2,7 +2,7 @@ const std = @import("std");
 const Builder = std.build.Builder;
 
 fn addFuzzStep(b: *Builder) !void {
-    const fuzz_lib = b.addStaticLibrary("fuzz-lib", "test/fuzz.zig");
+    const fuzz_lib = b.addStaticLibrary("fuzz-lib", "test/fuzz/fuzz_lib.zig");
     fuzz_lib.addPackagePath("aro", "src/lib.zig");
     fuzz_lib.setBuildMode(.Debug);
     fuzz_lib.want_lto = true;
@@ -11,8 +11,8 @@ fn addFuzzStep(b: *Builder) !void {
     const fuzz_executable_name = "fuzz";
     const fuzz_exe_path = try std.fs.path.join(b.allocator, &.{ b.cache_root, fuzz_executable_name });
 
-    // We want `afl-clang-lto -o path/to/output path/to/library`
-    const fuzz_compile = b.addSystemCommand(&.{ "afl-clang-lto", "-o", fuzz_exe_path });
+    // We want `afl-clang-lto -o path/to/output test/fuzz/main.c path/to/library`
+    const fuzz_compile = b.addSystemCommand(&.{ "afl-clang-lto", "-o", fuzz_exe_path, "test/fuzz/main.c" });
     // Add the path to the library file to afl-clang-lto's args
     fuzz_compile.addArtifactArg(fuzz_lib);
 
@@ -23,15 +23,6 @@ fn addFuzzStep(b: *Builder) !void {
     const fuzz_compile_run = b.step("fuzz", "Build executable for fuzz testing using afl-clang-lto");
     fuzz_compile_run.dependOn(&fuzz_compile.step);
     fuzz_compile_run.dependOn(&fuzz_install.step);
-
-    // Compile a companion exe for debugging crashes
-    const fuzz_debug_exe = b.addExecutable("fuzz-debug", "test/fuzz.zig");
-    fuzz_debug_exe.addPackagePath("aro", "src/lib.zig");
-    fuzz_debug_exe.setBuildMode(.Debug);
-
-    // Only install fuzz-debug when the fuzz step is run
-    const install_fuzz_debug_exe = b.addInstallArtifact(fuzz_debug_exe);
-    fuzz_compile_run.dependOn(&install_fuzz_debug_exe.step);
 }
 
 pub fn build(b: *Builder) !void {
